Update tests for Dashboard components
